Broke FrameDemo.java out into separate .java files for clarity and modularity.
