--- src/main/java/com/error404/project3/view/FrameDemo.java
@@ -1,202 +1,0 @@
-package com.error404.project3.view;
-
-import java.awt.*;
-import java.awt.event.*;
-import javax.swing.*;
-
-/* FrameDemo.java requires no other files. */
-public class FrameDemo {
-
-  private static void searchGui() {
-    // Create and set up main window.
-    JFrame frame = new JFrame("Search");
-    frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-
-    // Search label and text field
-    // ************************************************************* */
-    JLabel searchLabel = new JLabel("Search:");
-    searchLabel.setBounds(25, 50, 100, 30);
-    frame.add(searchLabel);
-
-    JTextField searchBox = new JTextField(""); // search box
-    searchBox.setBounds(100, 50, 350, 30);
-    frame.add(searchBox);
-
-    // Main results field
-    // ************************************************************* */
-    JTextArea searchReturn = new JTextArea(); // return text box
-    searchReturn.setBounds(25, 150, 550, 150);
-    // jScrollPane1.setViewportView(maintTextArea);
-    frame.add(searchReturn);
-
-    // Indexed files labels
-    // ************************************************************* */
-    JLabel indexLabel = new JLabel("Indexed Files: ");
-    indexLabel.setBounds(250, 350, 80, 30);
-    frame.add(indexLabel);
-
-    JLabel indexCount = new JLabel("0"); // indexed file return text
-    indexCount.setBounds(330, 350, 100, 30);
-    frame.add(indexCount);
-
-    // Radio buttons
-    // *************************************************************** */
-
-    JRadioButton allTermButton = new JRadioButton("All Terms");
-    allTermButton.setBounds(112, 100, 100, 50);
-    frame.add(allTermButton);
-    allTermButton.setSelected(true); // Set default selection
-
-    JRadioButton anyTermButton = new JRadioButton("Any Terms");
-    anyTermButton.setBounds(274, 100, 100, 50);
-    frame.add(anyTermButton);
-
-    JRadioButton exactTermButton = new JRadioButton("Exact Term");
-    exactTermButton.setBounds(436, 100, 100, 50);
-    frame.add(exactTermButton);
-
-    ButtonGroup buttons = new ButtonGroup();
-    buttons.add(allTermButton);
-    buttons.add(anyTermButton);
-    buttons.add(exactTermButton);
-
-    // About button
-    // ***************************************************************** */
-    JButton aboutButton = new JButton("About"); // About button
-    aboutButton.setBounds(450, 350, 125, 30);
-    frame.add(aboutButton);
-
-    aboutButton.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        aboutGui();
-      }
-    });
-
-    // Maintenance button
-    // *************************************************************** */
-    JButton maintainButton = new JButton("Maintenance"); // Maintenance button
-    maintainButton.setBounds(25, 350, 125, 30);
-    frame.add(maintainButton);
-
-    maintainButton.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        maintenanceGui();
-      }
-    });
-
-    // Search button
-    // *********************************************************************** */
-    JButton searchButton = new JButton("Search"); // Search Button
-    searchButton.setBounds(475, 50, 100, 30);
-    frame.add(searchButton);
-
-    searchButton.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        searchReturn.setText("Oops! Error 404: Not Found");
-      }
-    });
-
-    // Empty label to set window dimensions?
-    // ************************************************************************ */
-    JLabel emptyLabel = new JLabel("");
-    emptyLabel.setPreferredSize(new Dimension(600, 400)); // Sets main search window size
-    frame.getContentPane().add(emptyLabel, BorderLayout.CENTER);
-
-    // Display the window.
-    frame.pack();
-    frame.setVisible(true);
-    frame.setSize(625, 450);
-    frame.setLocationRelativeTo(null); 
-  }
-
-  private static void aboutGui() { // About window
-
-    // Create and set up the window.
-    JFrame frameAbout = new JFrame("About");
-
-    JLabel label = new JLabel("");
-    label.setPreferredSize(new Dimension(400, 400));
-
-    label.setText(
-        "<html>Developer names<br><br>Lauren Bartlett<p>SeanBrown<p>Briana McKenzie<br><br>Error404_Project3<br><br />https://github.com/DeltaWild/Error404_Project3.git"
-            + "<br /><br />I don't know HTML, obviously  </html>");
-
-    frameAbout.getContentPane().add(label, BorderLayout.CENTER);
-
-    // Display the window.
-    frameAbout.pack();
-    frameAbout.setVisible(true);
-  }
-
-  private static void maintenanceGui() {
-    // Create and set up the window.
-    JFrame frame = new JFrame("Maintenance");
-
-    JTextArea maintTextArea = new JTextArea(); // Maintenance return text area
-    maintTextArea.setBounds(25, 75, 550, 200);
-    // jScrollPane1.setViewportView(maintTextArea);
-    frame.add(maintTextArea);
-
-    // *********************************************************** */
-    JButton maintAddFile = new JButton("Add File"); // Add File button
-    maintAddFile.setBounds(37, 300, 150, 30);
-    frame.add(maintAddFile);
-
-    maintAddFile.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        maintTextArea.setText("404 (Not Found)");
-      }
-    });
-    // *************************************************************** */
-    JButton maintRebuild = new JButton("Rebuild Out of Date"); // Rebuild button
-    maintRebuild.setBounds(224, 300, 150, 30);
-    frame.add(maintRebuild);
-
-    maintRebuild.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        maintTextArea.setText("401 (Unauthorized)");
-      }
-    });
-    // *************************************************************** */
-    JButton maintRemoveFile = new JButton("Remove File"); // Remove File button
-    maintRemoveFile.setBounds(411, 300, 150, 30);
-    frame.add(maintRemoveFile);
-
-    maintRemoveFile.addActionListener(new ActionListener() {
-      public void actionPerformed(ActionEvent e) {
-        maintTextArea.setText("403 (Forbidden)");
-      }
-    });
-    // *************************************************************** */
-
-    JLabel maintLabel1 = new JLabel("Search Engine - Index Maintenance"); // Title
-    maintLabel1.setBounds(125, 25, 350, 30);
-    frame.add(maintLabel1);
-    maintLabel1.setFont(maintLabel1.getFont().deriveFont(20.0f));
-
-    JLabel maintLabel2 = new JLabel("Indexed Files: 0"); // Index file count
-    maintLabel2.setBounds(250, 350, 100, 30);
-    frame.add(maintLabel2);
-    // *********************************************************** */
-
-    JLabel emptyLabel = new JLabel("");
-    emptyLabel.setPreferredSize(new Dimension(600, 400)); // sets window size
-    frame.getContentPane().add(emptyLabel, BorderLayout.CENTER);
-
-    // Display the window.
-    frame.pack();
-    frame.setVisible(true);
-    frame.setLocationRelativeTo(null); 
-  }
-
-  public static void main(String[] args) {
-    // Schedule a job for the event-dispatching thread:
-    // creating and showing this application's GUI.
-    javax.swing.SwingUtilities.invokeLater(new Runnable() {
-      public void run() {
-        searchGui();
-
-      }
-    });
-  }
-}
