+++ src/main/java/com/error404/project3/view/Main.java
@@ -1,0 +1,14 @@
+package com.error404.project3.view;
+
+public class Main {
+    public static void main(String[] args) {
+        // Schedule a job for the event-dispatching thread:
+        // creating and showing this application's GUI.
+        javax.swing.SwingUtilities.invokeLater(new Runnable() {
+            public void run() {
+                SearchGUI.searchGui();
+
+            }
+        });
+    }
+}
